refactor: perform all changes recommended in the third review, including:
refactor: add exception handlers for non-handled errors
fix: find users method to comply with specification
refactor: simplify getForUserEvent method
feat: add statistics to private and admin services
refactor: exclude repository methods from loops
refactor: exclude repository methods from loops

+++ ewm-service/src/main/java/yandex/praktikum/ewmservice/controllers/admin/AdminUsersController.java
@@ -32,8 +32,8 @@ public class AdminUsersController {
     @GetMapping
     public List<UserDto> getAll(
             @RequestParam(name = "ids", required = false) List<Long> ids,
-            @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") long from,
-            @Positive @RequestParam(name = "size", defaultValue = "10") long size) {
+            @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") int from,
+            @Positive @RequestParam(name = "size", defaultValue = "10") int size) {
         return adminUsersService.getAll(ids, from, size);
     }
